feat: Añade todos los service y demás

+++ src/main/java/mathtexpedia/es/api/service/subject/SubjectServiceImpl.java
@@ -7,6 +7,7 @@ import mathtexpedia.es.api.domain.model.subject.CreateSubjectDto;
 import mathtexpedia.es.api.domain.model.subject.SubjectDto;
 import mathtexpedia.es.api.domain.model.subject.UpdateSubjectDto;
 import mathtexpedia.es.api.persistence.pdf.PDFDataService;
+import mathtexpedia.es.api.persistence.quiz.QuizDataService;
 import mathtexpedia.es.api.persistence.subject.Subject;
 import mathtexpedia.es.api.persistence.subject.SubjectDataService;
 import mathtexpedia.es.api.persistence.subjectUnit.SubjectUnitDataService;
@@ -25,11 +26,18 @@ public class SubjectServiceImpl implements SubjectService{
     private final SubjectDataService subjectDataService;
     private final SubjectUnitDataService subjectUnitDataService;
     private final PDFDataService pdfDataService;
+    private final QuizDataService quizDataService;
 
-    public SubjectServiceImpl(SubjectDataService subjectDataService, SubjectUnitDataService subjectUnitDataService, PDFDataService pdfDataService) {
+    public SubjectServiceImpl(
+            SubjectDataService subjectDataService,
+            SubjectUnitDataService subjectUnitDataService,
+            PDFDataService pdfDataService,
+            QuizDataService quizDataService
+    ) {
         this.subjectDataService = subjectDataService;
         this.subjectUnitDataService = subjectUnitDataService;
         this.pdfDataService = pdfDataService;
+        this.quizDataService = quizDataService;
     }
 
     @Override
@@ -72,13 +80,14 @@ public class SubjectServiceImpl implements SubjectService{
         Subject toDelete = subjectDataService.getById(id)
                 .orElseThrow(() -> new MathtexpediaNotFoundException("Subject not found with id: " + id));
 
-        if (!subjectUnitDataService.getAllForSubject(id).isEmpty()) {
+        if (!subjectUnitDataService.getAllForSubject(id).isEmpty())
             throw new MathtexpediaConflictException("Cannot delete subject with id: " + id + " because it has associated subject units.");
-        }
 
-        if (!pdfDataService.getGeneralPDFsForSubject(id).isEmpty()) {
+        if (!pdfDataService.getGeneralPDFsForSubject(id).isEmpty())
             throw new MathtexpediaConflictException("Cannot delete subject with id: " + id + " because it has associated PDFs.");
-        }
+
+        if (!quizDataService.getAllBySubjectId(id).isEmpty())
+            throw new MathtexpediaConflictException("Cannot delete subject with id: " + id + " because it has associated quizzes.");
 
         subjectDataService.delete(toDelete);
     }
